Completed the allocate domain service. Rearranged some code in files. Added tests for allocate domain service.
