+++ model.py
@@ -44,3 +44,16 @@ class Batch:
     if other.eta is None:
       return True
     return self.eta > other.eta
+
+class OutOfStock(Exception):
+  pass
+
+def allocate(line: OrderLine, batches: List[Batch]) -> str:
+  try:
+    batch = next(
+      b for b in sorted(batches) if b.can_allocate(line)
+    )
+    batch.allocate(line)
+    return batch.reference
+  except StopIteration:
+    raise OutOfStock(f'Out of stock for sku{line.sku}')
