+++ test_batches.py
@@ -1,7 +1,7 @@
-from model import Batch, OrderLine
+from model import Batch, OrderLine, allocate, OutOfStock
 from datetime import date
 import datetime
-from services import allocate
+import pytest
 
 def test_allocating_to_a_batch_reduces_the_available_quantity():
   batch = Batch("batch-001", "SMALL-TABLE", qty=20, eta=date.today())
@@ -76,3 +76,10 @@ def test_returns_allocated_batch_ref():
   line = OrderLine("oref", "HIGHBROW-POSTER", 10)
   allocation = allocate(line, [in_stock_batch, shipment_batch])
   assert allocation == in_stock_batch.reference
+
+def test_raises_out_of_stock_exception_if_cannot_allocate():
+  batch = Batch('batch1', 'SMALL-FORK', 10, eta=today)
+  allocate(OrderLine('order1','SMALL-FORK',10),[batch])
+
+  with pytest.raises(OutOfStock, match='SMALL-FORK'):
+    allocate(OrderLine('order2','SMALL-FORK', 1), [batch])
